Remove flag btsec_check_valid_discovery_database
This flag has been soaked for more than a month and is ready for
removal.

This is a cherry-pick of ag/30363401 -- the 25Q2 branch was cut with the flag still present and we need to remove that flag to close.

Bug: 356201480
Flag: EXEMPT flag removal
Test: m libbluetooth
Tag: #security
Ignore-AOSP-First: Security
Change-Id: I31d47316748515fb44d729ba14946f98420afb36
(cherry picked from commit 769caf391c6055c6f9db945b71d96b2f01c8799c)